+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # Developing
 
+- Allow revoke refresh token.
+- Add logout.
 - Register.
 - Authentication.
 - Access token and refresh token.